mempool: support alignment <= page_size in mempool allocator
When a memory block is requested with alignment requirement <= page_size
it is sufficient to return an object that is the first one in a mempool
page. However, since initialization of a new mempool page is expensive
it is not performed at each allocation with custom alignment requirement,
but deferred to the moment when another objects from that page needs to
be allocated.

Signed-off-by: Paweł Dziepak <pdziepak@quarnos.org>